Changed AAD & Packing to Use Unsigned Integers
Changed the Albedo Ambient Diffuse buffer to store unsigned integers
rather than floats, and also changed the packing algorithm to pack and
unpack unsigned integers rather than floats. This dramatically helped
increase the precision of the values after they have been packed and
unpacked from the AAD Buffer, and also eliminated the need for
converting unsigned integers to floats.

+++ Honeycomb GE/res/shaders/render/deferred/light/ambientLightFS.glsl
@@ -18,7 +18,7 @@ void main() {
 							gl_FragCoord.y / camera.height);
 
 	vec3 aad = texture2D(gBufferAlbedoAmbientDiffuse, screenCoord).xyz;
-	vec3 ambient = unpackRGB(aad.g);
+	vec3 ambient = unpackRGB(uint(aad.g));
 
 	fragColor = vec4(calculateAmbientLight(ambientLight) * ambient, 1.0F);
 }
